# Batch report calls in hash_redis_store