--- a/lua/install.sh
+++ b/lua/install.sh
@@ -5,6 +5,6 @@ pip install hererocks
 hererocks $LUA_DIR -l5.1 -rlatest
 source $LUA_DIR/bin/activate
 lua -v
-for i in luacov busted redis-lua inspect; do 
+for i in luacov busted redis-lua inspect lua-cjson; do 
   luarocks install $i;
 done