--- a/lua/test/bootstrap.lua
+++ b/lua/test/bootstrap.lua
@@ -1,4 +1,5 @@
 inspect = require "inspect"
+cjson = require 'cjson'
 require './lua/test/redis-call'