Little bump up in cpu but much faster:

### Before:

![image](https://user-images.githubusercontent.com/96786/66276527-06cc4480-e862-11e9-8f0b-faa221ec3c2c.png)

```
➜  coverband git:(change_to_hash) COVERBAND_HASH_REDIS_STORE=t bundle exec rake benchmarks:redis_reporting
/Users/karlbaum/workspace/coverband/lib/coverband.rb:35: warning: already initialized constant Coverband::Adapters::RedisStore
/Users/karlbaum/workspace/coverband/lib/coverband/adapters/redis_store.rb:8: warning: previous definition of RedisStore was here
D, [2019-10-06T17:51:00.626595 #70722] DEBUG -- : using default configuration
runs benchmarks on reporting large sets of files to redis
Warming up --------------------------------------
   store_reports_all     1.000  i/100ms
Calculating -------------------------------------
   store_reports_all      2.029  (± 0.0%) i/s -     31.000  in  15.384839s
Warming up --------------------------------------
store_reports_subset     5.000  i/100ms
Calculating -------------------------------------
store_reports_subset     58.196  (± 8.6%) i/s -      1.155k in  20.025083s
```

### After:

![i mage](https://user-images.githubusercontent.com/96786/66276501-b9e86e00-e861-11e9-81f3-0d805c7f39db.png)

```
➜  coverband git:(batch_call_lua) COVERBAND_HASH_REDIS_STORE=t bundle exec rake benchmarks:redis_reporting
runs benchmarks on reporting large sets of files to redis
Warming up --------------------------------------
   store_reports_all     1.000  i/100ms
Calculating -------------------------------------
   store_reports_all      4.660  (± 0.0%) i/s -     70.000  in  15.066954s
Warming up --------------------------------------
store_reports_subset    13.000  i/100ms
Calculating -------------------------------------
store_reports_subset    131.136  (± 6.9%) i/s -      2.613k in  20.049154s
```

Looks to be about 2X as fast according to our performance benchmark task.