@kbaum 2019-10-07
@danmayer  - to be clear, the cpu usage on redis still goes down about 20-25% compared to master for our benchmarks. I was comparing this iteration with the change_to_hash branch.